requirements.txt
jax==0.4.23
jaxlib==0.4.23
ml-dtypes==0.3.2
numpy==1.26.3
opt-einsum==3.3.0
PyYAML==6.0.1
scipy==1.12.0
